Move marshalTx/unmarshalTx to BCP
