--- packages/sdk38/types/decoding.d.ts
@@ -1,2 +1,0 @@
-import { StdTx } from "./types";
-export declare function unmarshalTx(data: Uint8Array): StdTx;
